Put TBAA tags on some loads and store.
There are some kludges here which will go away when some changes are
made to the way that types are defined.

The point of this change is so that the following test case will
generate only one load since it'll know that the store can't conflict.

package Tbaa is
   type I1 is range 1 .. 1000;
   type I2 is range 1 .. 10_000;

   function T (P1: access I1; P2: access I2) return Integer;
end Tbaa;
package body Tbaa is
   function T (P1: access I1; P2: access I2) return Integer is
      X, Y : Integer;
   begin
      X := Integer (P1.all);
      P2.all := 77;
      Y := Integer (P1.all);
      return X + Y;
   end T;
end Tbaa;