Properly handle Packed_Array_Impl_Type
There are a number of issues here, including missing some cases by checking
too late (e.g., where we first check for discrete types, then arrays).
But also, a Packed_Array_Impl_Type isn't always scalar, but can itself
be an array.

Doing it in this way isn't the cleanest approach, but is part of a larger
problem: in Gigi, we quickly convert GNAT types to GCC types and then do
tests as to what we have (record, array, etc) on the GCC type data.  Here,
we do the checking each time on the GNAT tree, which means that complex
tests (such as this one or handling private types) need to be done at
each use.  That's messy and error-prone, so some new mechanism needs to
be found.  For later.

Fixes cd3021a regression.