Restructure what's in some specs
The main purpose of this change is to put all items about GLValues
into GNATLLVM.GLValue, rather than having parts of it in
GNATLLVM.Environment.  This was impossible before moving things out
of the Env structure, but is still tricky to avoid having "with"
circularity among some specs, particularly GNATLLVM.GLValues,
GNATLLVM.Utils, and GNATLLVM.Environment.  This circularity was
broken by the addition of GNATLLVM.Core.