Support Is_Constr_Subt_For_UN_Aliased
Also support saving address from N_Attribute_Definition_Clause and
make many infrastruction changes/improvements for relationships.

NOTE: This is a kludge half-and-half implementation and not complete support
for allocators.  This is being checked in to create a good break point for
further work.

NOTE: There are also ACATS regressions that will be fixed later, once
the implementation change has been completed.