Add support for variant records
This adds support for tracking the structure of variant records.
It's missing code to compute the per-record size based on the value of
the discriminant, currently just computing the maximum size.

This fixes c54a07a and three XFAIL'ed tests, c34006d, c34007p, and
c34007r.

Unfortunately, it now notices that it can't properly support a case
that we previously supported "by accident", which was a tagged variant
record with an extension.  We lay out the parent fields as part of our
record, so we end up trying to put records after the variant part,
which is not something envisioned by the current data structure, though
there are some possible extensions to it that will support it.  Two
of these tests (c390010 and cc50a01) are technically "regressions",
though they never worked correctly and the second of those was "fixed"
three days ago.