Restore -enable-shrink-wrap=0
It "works", in the sense that it doesn't produce clearly erroneous
code, but it can end up in a situation where it produces references to
variable in the after cutting back the stack frame.  That likely
works nowadays (but there's a potential concern about signal handlers),
but isn't a great idea and valgrind complains about it.