Change Is_Loadable_Type to count in alignment units and related cleanups
We care how many instructions are needed to move to and from an object
and that's the size divided by the alignment, so change to that.

Making that change exposed some bugs related to using the Relationship
of Object instead of Data when parameter passing is done, so those
were fixed.  That work suggested three cleanups:

(1) Use more named parameters in Emit_Assignment
(2) Remove "LLVM" from names of type size functions and use overloading
    instead.
(3) Remove "Has_" functions and replace with use of No and Present on Get_.