Add facility to convert composite constants to new type
We could already convert constant structs to a type with identical layouts
by rebuilding the constant.  This change addes a feature where we can do
the equivalent of an unchecked conversion of a constant to some other type.
We do this by exploting the fact that the LLVM optimizer knows how to undo
pointer-punning of a constant to or from an array of bytes.

In addition to adding the subprograms that do this conversion, we also
modify the tests to determine which expressions can be evaluated statically
to take advantage of this new capability and fix some latent bugs that
these changes made visible.