Abandon idea of RI-specific TBAA tags
The original idea here was to assign a TBAA tag to an RI that's part of a
variable-sized record.  But this can't work in the presence of subtypes
because a subtype might be a fixed-size record whose single RI has as a
parent multiple RI's from the base type and there's no way to represent
that.  So we don't attempt to assign TBAA tags to those RI's.  We can
and will directly pick up the TBAA type of the field instead.  The number
of cases where we load a full RI's worth of data is very small, so the
loss here is small as well.

Change-Id: Iba111e94e5d2e8d93d7fb44f12f1891a8f7d9041