Major cleanup of zero-length objects and types
The goal of this work is to remove errors and warnings produced by GCC
and clang relating to arrays of size 0. If an array is of dynamic size
and ends up zero, we have no issue. The problem is arrays that are
statically of size zero. Such arrays aren't very common in user code
(mostly null strings) but are used a lot in ACATS tests.

We handle this in a number of ways:

(1) We change GNAT LLVM to ignore lots of things relating to arrays of zero
size when emitting C. For example, loads of such objects are forced to
Undef and stores are suppressed. We remove returns and parameters of
subprograms that are of zero size.

(2) We omit fields of zero size unless they're at the end of a struct
and we're using C99 or later.

(3) For versions of C earlier than C99, we change all references of
a zero-length array to an array of length 1, for example in pointers
to 0-length arrays.  For C99 and later, we use the indefinite ("[]")
syntax.

Change-Id: I5ae5cdaf43bd1ecc791d70740b4e1a47cdfe4f0d