Restore 80-bit extended precision on x86_64
It turns out that LLVM uses the term "precision" for the number of bits
in the significand, whereas GNAT uses the same term to refer to the
total number of bits. (Note that this is still different from the
storage size; for example, 80-bit floats on x86_64 have 64 significant
bits, 80 bits overall, and require 128 bits of storage.) By using LLVM's
"size in bits", we match GNAT's "precision", thus making it use 80-bit
extended precision on x86_64 again.

TN: W116-012