Refine size computation for move
The last change to Compute_Size added more cases where we used the RHS
values, but when setting the No_Padding value, we always used the LHS.
So we now record which size we need and use all the values (including
whether a size is unconstrained) to compute the size.  This fixes
a regression.

TN: TB23-035