Support address conversions with tagged pointers
This commit adds support for conversions between different address
types, as well as to and from integers, for tagged-pointer targets.
Since we store addresses as pointers at the LLVM level, conversions
between address types are a no-op; conversion to integer extracts the
address; and conversion from integer derives a tagged pointer from the
null pointer.

Part of eng/toolchain/gnat-llvm#49.