Represent address types as pointers
While our runtime treats System.Address and friends as modular types, we
need to represent the values as pointers at the LLVM level when
targeting an architecture with tagged pointers; otherwise, we would lose
the ability to convert them back to valid pointers. For address
literals, we don't have provenance information, so we derive a
capability from the null pointer (which can't be dereferenced).

Part of eng/toolchain/gnat-llvm#49.