Get the declarations of intrinsics from LLVM
Instead of manually declaring LLVM intrinsics, we now ask LLVM to emit
the correct declaration; it can do so even for overloaded intrinsics.
The advantage is that we don't have to worry about name mangling, and
that LLVM automatically adds the right function attributes for each
intrinsic.

Part of eng/toolchain/gnat-llvm#49.