Support static address clauses with tagged pointers
Our usual approach to address clauses with static addresses is to
replace the global object with a pointer that we create from the
user-provided address. On tagged-pointer architectures, we can't do this
because pointers can't be created from arbitrary integers; instead, we
rely on a custom form of the "absolute_symbol" metadata annotation in
LLVM bitcode.

Fixes eng/toolchain/gnat-llvm#77.