More work on inline and CCG
For most instances of inlined functions, just write "inline" and
possibly "static" in both declaration and definition. If we're defining
an inline function in the main unit, add "extern" to the definition. This
is the C99 way of handling inline and GCC supports it since 4.3.

Because versions between C90 and C99 support "inline", but not necessarily
the above behavior, don't support inter-unit inlining for those versions
unless -O2 is specified.

If we've made a variable for a constant that was used in a non-inlined
function and now it's being used in an inline function, make it global.

Scan for other constants that are used in inlined functions so we don't
make them static and we qualify them with our input filename.