Use libunwind for stage 2 and 3 builds
We're using the libunwind headers during our build, so the runtime
depends on our libunwind (which is an unfortunate side effect of the
unwind-pe.h file that we added to libunwind). Before we dropped
--target=llvm, stages 2 and 3 were built with Clang; restore this
behavior by putting the path to Clang first in the search path, causing
gprbuild to prefer Clang over GCC. This gives us back the ability to
link with libunwind and thus fixes the bootstrap build.

Issue: eng/toolchain/gnat-llvm#149