Fix the plan for fixedbugs testing
Since this testing depends on aunit, gnatcoll-core, libgpr and xmlada,
we need to rebuild those libs (manually). In theory, this should be
implied by the fact that the gnat-llvm repository is 'anod
vcs'ed. However, since we manually install gnat-llvm in the plan
(since it has been built in the previous job), those libs are not
detected as being in need of a rebuild.

Another solution would have been to launch their build directly
in the 'build' job. However, since those are needed only for fixedbugs
(and not for acats or ccg), it's conceptually cleaner to perform
them only in the fixedbugs job.