Do not sign-extend unsigned bounds in DWARF
The gdb test array_char_idx.exp was failing.  For this source code:

   type Char_Table is array (Character range Character'First .. Character'Last)
      of Natural;

... the resulting DWARF was:

 <1><ea6>: Abbrev Number: 5 (DW_TAG_array_type)
    <ea7>   DW_AT_type        : <0xeb7>
    <eab>   DW_AT_name        : (indirect string, offset: 0x14a9): pck__char_table
 <2><eaf>: Abbrev Number: 6 (DW_TAG_subrange_type)
    <eb0>   DW_AT_type        : <0xed1>
    <eb4>   DW_AT_lower_bound : 0
    <eb5>   DW_AT_upper_bound : -1

What happened here is that the upper bound of 255 is sign-extended by
UI_To_LLVM.

I believe this occurs because LLVM's scalar types are all signed (with
unsigned being handled by unsigned operations).  However, DWARF does
not work the same way, and so this patch undoes the sign-extension in
this case.