Use "git am"-style patches for LLVM DWARF changes
The LLVM patches to DWARF generation are getting a bit unwieldy.
First, I forgot to add some prerequisite patches.  Second, because the
patches tend to touch the same code, there is only a certain order in
which they can be applied.

To solve these problems, I made a new local branch of LLVM, applied
all the DWARF patches, and then rebased to reorder and squash the
patches so that each one is more or less a single "topic".  Then I
used "git format-patch" to produce the patches, so that the ordering
is obvious from the file name.