Phase one of opaque pointer work
This is for #31.

This commit create a new framework, called MD type, which are at a similar
semantic level to LLVM types except that they contain information about
signedness, always show what a pointer's designated type is, and includes
field name information for structs.

This change makes most of GNAT LLVM create MD_Types instead of LLVM types
directly and points GL_Types to MD_Types.  The bulk of the generation
of actual LLVM types is from MD_Types.

This keeps in place the current mechanism of transmitting signedness
data to CCG and changes nothing in terms of functionality (though
there are some minor output changes relating to names of fields for
padding and bitfields).