Don't set the nocapture attribute for function parameters
We've seen a miscompilation where a procedure takes two references and
stores one in a field of the record referenced by the other. In general,
I don't see a way of guaranteeing that a procedure doesn't store a
reference parameter (for example, in a global).

This is really only relevant for functions which are defined in another
compilation unit; LLVM should be able to infer nocapture for local
functions.

Fixes eng/toolchain/gnat-llvm#244.