Emit DWARF for non-constant field offsets
This changes gnat-llvm to emit debuginfo for records whose fields have
non-constant field offsets or sizes.  This also handles the case where
an array has a length that is determined by a discriminant.

The general approach is to turn back annotation expressions into DWARF
expressions.  Some notes:

* The various forms of div and mod are treated identically, which is
  probably erroneous; however I wasn't sure whether these can all
  actually be seen here, and at the same time this is poorly specified
  in DWARF.  See this email (and a similar thread from the previous
  month):

    https://lists.dwarfstd.org/pipermail/dwarf-discuss/2025-October/002736.html

* References to variables are not handled yet.  This is intentional as
  I think this is not straightforward and felt that this patch was
  already big enough.

For array types whose length comes from a discriminant, the array now
uses the enclosing record as its scope, and the subrange refers
directly to the discriminant field (when LLVM allows this).