Handle inherited discriminants in DWARF
gnat-llvm crashed while generating debuginfo for a variant part where
the corresponding discriminant was inherited from a parent.

It's not completely clear how this should be represented in DWARF, as
I think it doesn't really make sense to refer to a DIE in some other
type (and anyway implementing that might be difficult in LLVM).

So, I decided to emit such discriminants in the derived type, but
marked DW_AT_artificial.  I've also patched gdb to ignore such fields
when printing (but not when computing the runtime type).