Select a default CPU for x86 as Clang does
For x86 triples and if not selected explicitly, set the same default
CPU as Clang does. Consistency with Clang is generally useful, and on
x86 in particular, the choice of CPU results in selection of features.
In this case the use of SSE rather than x87 for floating point
operations improve rounding behavior.

issue: eng/toolchain/gnat-llvm#285